Поправил название переменных, сделал комментарии, убрал лишние отступы, привел классы в порядок
