--- src/main/java/Belousov/Main.java
@@ -1,9 +1,0 @@
-package Belousov;
-
-public class Main {
-    public static void main(String[] args) {
-
-        Service service = new Service();
-        service.service();
-    }
-}
