+++ src/main/java/com/epam/belousov/Main.java
@@ -1,0 +1,10 @@
+package com.epam.belousov;
+
+import com.epam.belousov.service.Service;
+
+public class Main {
+    public static void main(String[] args) {
+        Service service = new Service();
+        service.service();
+    }
+}
